+++ testing/label_service.go
@@ -115,7 +115,7 @@ func CreateLabel(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "label_1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Properties: map[string]string{
 							"color": "fff000",
 						},
@@ -126,7 +126,7 @@ func CreateLabel(
 				label: &influxdb.Label{
 					ID:    MustIDBase16(labelTwoID),
 					Name:  "label_1",
-					OrgID: MustIDBase16(orgOneID),
+					OrgID: idOne,
 					Properties: map[string]string{
 						"color": "fff000",
 					},
@@ -137,7 +137,7 @@ func CreateLabel(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "label_1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Properties: map[string]string{
 							"color": "fff000",
 						},
@@ -158,7 +158,7 @@ func CreateLabel(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "tag_1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Properties: map[string]string{
 							"color": "fff000",
 						},
@@ -169,7 +169,7 @@ func CreateLabel(
 				label: &influxdb.Label{
 					ID:    MustIDBase16(labelOneID),
 					Name:  "     tag_1     ",
-					OrgID: MustIDBase16(orgOneID),
+					OrgID: idOne,
 					Properties: map[string]string{
 						"color": "fff000",
 					},
@@ -180,7 +180,7 @@ func CreateLabel(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "tag_1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Properties: map[string]string{
 							"color": "fff000",
 						},
@@ -201,7 +201,7 @@ func CreateLabel(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "tag_1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Properties: map[string]string{
 							"color": "fff000",
 						},
@@ -212,7 +212,7 @@ func CreateLabel(
 				label: &influxdb.Label{
 					ID:    MustIDBase16(labelOneID),
 					Name:  "TAG_1",
-					OrgID: MustIDBase16(orgOneID),
+					OrgID: idOne,
 					Properties: map[string]string{
 						"color": "fff000",
 					},
@@ -223,7 +223,7 @@ func CreateLabel(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "tag_1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Properties: map[string]string{
 							"color": "fff000",
 						},
@@ -246,7 +246,7 @@ func CreateLabel(
 				label: &influxdb.Label{
 					Name:  "Tag2",
 					ID:    MustIDBase16(labelOneID),
-					OrgID: MustIDBase16(orgOneID),
+					OrgID: idOne,
 					Properties: map[string]string{
 						"color": "fff000",
 					},
@@ -257,7 +257,7 @@ func CreateLabel(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "Tag2",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Properties: map[string]string{
 							"color": "fff000",
 						},
@@ -313,12 +313,12 @@ func FindLabels(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "Tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
 						Name:  "Tag2",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 				},
 			},
@@ -330,12 +330,12 @@ func FindLabels(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "Tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
 						Name:  "Tag2",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 				},
 			},
@@ -347,24 +347,24 @@ func FindLabels(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "Tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
 						Name:  "Tag2",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 					{
 						ID:    MustIDBase16(labelThreeID),
 						Name:  "Tag1",
-						OrgID: MustIDBase16(orgTwoID),
+						OrgID: idTwo,
 					},
 				},
 			},
 			args: args{
 				filter: influxdb.LabelFilter{
 					Name:  "Tag1",
-					OrgID: idPtr(MustIDBase16(orgOneID)),
+					OrgID: idPtr(idOne),
 				},
 			},
 			wants: wants{
@@ -372,7 +372,7 @@ func FindLabels(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "Tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 				},
 			},
@@ -384,14 +384,14 @@ func FindLabels(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 				},
 			},
 			args: args{
 				filter: influxdb.LabelFilter{
 					Name:  "TAG1",
-					OrgID: idPtr(MustIDBase16(orgOneID)),
+					OrgID: idPtr(idOne),
 				},
 			},
 			wants: wants{
@@ -399,7 +399,7 @@ func FindLabels(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 				},
 			},
@@ -446,12 +446,12 @@ func FindLabelByID(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "Tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
 						Name:  "Tag2",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 				},
 			},
@@ -462,7 +462,7 @@ func FindLabelByID(
 				label: &influxdb.Label{
 					ID:    MustIDBase16(labelOneID),
 					Name:  "Tag1",
-					OrgID: MustIDBase16(orgOneID),
+					OrgID: idOne,
 				},
 			},
 		},
@@ -524,7 +524,7 @@ func UpdateLabel(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 					},
 				},
@@ -539,7 +539,7 @@ func UpdateLabel(
 				labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "NotTag1",
 					},
 				},
@@ -551,12 +551,12 @@ func UpdateLabel(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "tag_1",
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "tag_2",
 					},
 				},
@@ -571,12 +571,12 @@ func UpdateLabel(
 				labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "tag_1",
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "tag_2",
 					},
 				},
@@ -593,12 +593,12 @@ func UpdateLabel(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "tag_1",
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "tag_2",
 					},
 				},
@@ -613,12 +613,12 @@ func UpdateLabel(
 				labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "tag_1",
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "tag_2",
 					},
 				},
@@ -635,7 +635,7 @@ func UpdateLabel(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 					},
 				},
@@ -652,7 +652,7 @@ func UpdateLabel(
 				labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 						Properties: map[string]string{
 							"color": "fff000",
@@ -667,7 +667,7 @@ func UpdateLabel(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 						Properties: map[string]string{
 							"color":       "fff000",
@@ -688,7 +688,7 @@ func UpdateLabel(
 				labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 						Properties: map[string]string{
 							"color":       "abc123",
@@ -704,7 +704,7 @@ func UpdateLabel(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 						Properties: map[string]string{
 							"color":       "fff000",
@@ -725,7 +725,7 @@ func UpdateLabel(
 				labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 						Properties: map[string]string{
 							"color": "fff000",
@@ -801,12 +801,12 @@ func DeleteLabel(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 					},
 					{
 						ID:    MustIDBase16(labelTwoID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag2",
 					},
 				},
@@ -818,7 +818,7 @@ func DeleteLabel(
 				labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelTwoID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag2",
 					},
 				},
@@ -830,7 +830,7 @@ func DeleteLabel(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 					},
 				},
@@ -842,7 +842,7 @@ func DeleteLabel(
 				labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 					},
 				},
@@ -899,7 +899,7 @@ func CreateLabelMapping(
 				Labels: []*influxdb.Label{
 					{
 						ID:    MustIDBase16(labelOneID),
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 						Name:  "Tag1",
 					},
 				},
@@ -907,10 +907,10 @@ func CreateLabelMapping(
 			args: args{
 				mapping: &influxdb.LabelMapping{
 					LabelID:    MustIDBase16(labelOneID),
-					ResourceID: MustIDBase16(bucketOneID),
+					ResourceID: idOne,
 				},
 				filter: &influxdb.LabelMappingFilter{
-					ResourceID: MustIDBase16(bucketOneID),
+					ResourceID: idOne,
 				},
 			},
 			wants: wants{
@@ -918,7 +918,7 @@ func CreateLabelMapping(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "Tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 				},
 			},
@@ -932,7 +932,7 @@ func CreateLabelMapping(
 			args: args{
 				mapping: &influxdb.LabelMapping{
 					LabelID:    MustIDBase16(labelOneID),
-					ResourceID: MustIDBase16(bucketOneID),
+					ResourceID: idOne,
 				},
 			},
 			wants: wants{
@@ -996,23 +996,23 @@ func DeleteLabelMapping(
 					{
 						ID:    MustIDBase16(labelOneID),
 						Name:  "Tag1",
-						OrgID: MustIDBase16(orgOneID),
+						OrgID: idOne,
 					},
 				},
 				Mappings: []*influxdb.LabelMapping{
 					{
 						LabelID:    MustIDBase16(labelOneID),
-						ResourceID: MustIDBase16(bucketOneID),
+						ResourceID: idOne,
 					},
 				},
 			},
 			args: args{
 				mapping: &influxdb.LabelMapping{
 					LabelID:    MustIDBase16(labelOneID),
-					ResourceID: MustIDBase16(bucketOneID),
+					ResourceID: idOne,
 				},
 				filter: influxdb.LabelMappingFilter{
-					ResourceID: MustIDBase16(bucketOneID),
+					ResourceID: idOne,
 				},
 			},
 			wants: wants{