+++ ui/src/perf/actions/index.ts
@@ -1,21 +1,14 @@
-export const SET_RENDER_ID = 'SET_RENDER_ID'
 export const SET_SCROLL = 'SET_SCROLL'
 export const SET_CELL_MOUNT = 'SET_CELL_MOUNT'
+export const SET_DASHBOARD_VISIT = 'SET_DASHBOARD_VISIT'
 
 export type Action =
-  | ReturnType<typeof setRenderID>
   | ReturnType<typeof setScroll>
   | ReturnType<typeof setCellMount>
+  | ReturnType<typeof setDashboardVisit>
 
 export type ComponentKey = 'dashboard'
 export type ScrollState = 'not scrolled' | 'scrolled'
-
-export const setRenderID = (component: ComponentKey, renderID: string) =>
-  ({
-    type: SET_RENDER_ID,
-    component,
-    renderID,
-  } as const)
 
 export const setScroll = (component: ComponentKey, scroll: ScrollState) =>
   ({
@@ -29,4 +22,11 @@ export const setCellMount = (cellID: string, mountStartMs: number) =>
     type: SET_CELL_MOUNT,
     cellID,
     mountStartMs,
+  } as const)
+
+export const setDashboardVisit = (dashboardID: string, startVisitMs: number) =>
+  ({
+    type: SET_DASHBOARD_VISIT,
+    dashboardID,
+    startVisitMs,
   } as const)