+++ ui/src/timeMachine/reducers/index.test.ts
@@ -71,7 +71,7 @@ describe('the Time Machine reducer', () => {
 
     it('builds the query if the query builder config is valid', () => {
       const expectedText =
-        'from(bucket: "metrics")\n  |> range(start: v.timeRangeStart, stop: v.timeRangeStop)\n  |> filter(fn: (r) => r["_measurement"] == "mem")\n  |> filter(fn: (r) => r["_field"] == "active")\n  |> aggregateWindow(every: v.windowPeriod, fn: mean)\n  |> yield(name: "mean")'
+        'from(bucket: "metrics")\n  |> range(start: v.timeRangeStart, stop: v.timeRangeStop)\n  |> filter(fn: (r) => r["_measurement"] == "mem")\n  |> filter(fn: (r) => r["_field"] == "active")\n  |> aggregateWindow(every: v.windowPeriod, fn: mean, createEmpty: false)\n  |> yield(name: "mean")'
       const validDraftQuery = {
         text: '',
         editMode: 'advanced',
@@ -102,6 +102,7 @@ describe('the Time Machine reducer', () => {
           ],
           aggregateWindow: {
             period: 'auto',
+            fillValues: false,
           },
         },
         hidden: false,