+++ ui/src/timeMachine/selectors/index.ts
@@ -1,30 +1,37 @@
 // Libraries
 import memoizeOne from 'memoize-one'
 import moment from 'moment'
-import {get, flatMap} from 'lodash'
+import {get} from 'lodash'
 import {fromFlux, Table} from '@influxdata/giraffe'
 
 // Utils
-import {parseResponse} from 'src/shared/parsing/flux/response'
 import {
   defaultXColumn,
   defaultYColumn,
+  mosaicYcolumn,
   getNumericColumns as getNumericColumnsUtil,
   getGroupableColumns as getGroupableColumnsUtil,
+  getStringColumns as getStringColumnsUtil,
 } from 'src/shared/utils/vis'
-import {getAllVariables, asAssignment} from 'src/variables/selectors'
-import {getWindowPeriod} from 'src/variables/utils/getWindowVars'
+import {
+  getWindowPeriod,
+  calcWindowPeriodForDuration,
+} from 'src/variables/utils/getWindowVars'
 import {
   timeRangeToDuration,
   parseDuration,
   durationToMilliseconds,
+  millisecondsToDuration,
 } from 'src/shared/utils/duration'
+
+//Selectors
+import {getAllVariables, asAssignment} from 'src/variables/selectors'
+import {getTimeRange} from 'src/dashboards/selectors'
 
 // Types
 import {
   QueryView,
   DashboardQuery,
-  FluxTable,
   AppState,
   DashboardDraftQuery,
   TimeRange,
@@ -68,12 +75,22 @@ export const getActiveWindowPeriod = (state: AppState) => {
   return getWindowPeriod(text, variables)
 }
 
-const getTablesMemoized = memoizeOne((files: string[]): FluxTable[] =>
-  files ? flatMap(files, parseResponse) : []
-)
+export const getWindowPeriodFromTimeRange = (state: AppState): string => {
+  const timeRange = getTimeRange(state)
+  if (timeRange.type === 'selectable-duration') {
+    return millisecondsToDuration(timeRange.windowPeriod)
+  }
 
-export const getTables = (state: AppState): FluxTable[] =>
-  getTablesMemoized(getActiveTimeMachine(state).queryResults.files)
+  if (timeRange.type === 'custom') {
+    const upper = Date.parse(timeRange.upper)
+    const lower = Date.parse(timeRange.lower)
+    return millisecondsToDuration(calcWindowPeriodForDuration(upper - lower))
+  }
+
+  throw new Error(
+    'Unknown timeRange type provided to getWindowPeriodFromTimeRange'
+  )
+}
 
 const getVisTableMemoized = memoizeOne(fromFlux)
 
@@ -87,11 +104,18 @@ export const getVisTable = (
 }
 
 const getNumericColumnsMemoized = memoizeOne(getNumericColumnsUtil)
+const getStringColumnsMemoized = memoizeOne(getStringColumnsUtil)
 
 export const getNumericColumns = (state: AppState): string[] => {
   const {table} = getVisTable(state)
 
   return getNumericColumnsMemoized(table)
+}
+
+export const getStringColumns = (state: AppState): string[] => {
+  const {table} = getVisTable(state)
+
+  return getStringColumnsMemoized(table)
 }
 
 const getGroupableColumnsMemoized = memoizeOne(getGroupableColumnsUtil)
@@ -114,7 +138,19 @@ export const getXColumnSelection = (state: AppState): string => {
 
 export const getYColumnSelection = (state: AppState): string => {
   const {table} = getVisTable(state)
-  const preferredYColumnKey = get(
+  const tm = getActiveTimeMachine(state)
+  let preferredYColumnKey
+
+  if (tm.view.properties.type === 'mosaic') {
+    preferredYColumnKey = get(
+      getActiveTimeMachine(state),
+      'view.properties.ySeriesColumns[0]'
+    )
+
+    return mosaicYcolumn(table, preferredYColumnKey)
+  }
+
+  preferredYColumnKey = get(
     getActiveTimeMachine(state),
     'view.properties.yColumn'
   )
@@ -141,12 +177,36 @@ const getSymbolColumnsSelectionMemoized = memoizeOne(
 )
 
 export const getFillColumnsSelection = (state: AppState): string[] => {
-  const validFillColumns = getGroupableColumns(state)
+  const {table} = getVisTable(state)
+  const tm = getActiveTimeMachine(state)
+  const graphType = tm.view.properties.type
+  let validFillColumns
+  if (graphType === 'mosaic') {
+    validFillColumns = getStringColumnsMemoized(table)
+  } else {
+    validFillColumns = getGroupableColumns(state)
+  }
 
   const preference = get(
     getActiveTimeMachine(state),
     'view.properties.fillColumns'
   )
+
+  if (graphType === 'mosaic') {
+    //user hasn't selected a fill column yet
+    if (preference === null) {
+      //check if value is a string[]
+      for (const key of validFillColumns) {
+        if (key.startsWith('_value')) {
+          return [key]
+        }
+      }
+      //check if value is a numeric column
+      if (table.columnKeys.includes('_value')) {
+        return []
+      }
+    }
+  }
 
   const {fluxGroupKeyUnion} = getVisTable(state)
 
@@ -237,6 +297,17 @@ export const getSaveableView = (state: AppState): QueryView & {id?: string} => {
   }
 
   // TODO: remove all of these conditionals
+  if (saveableView.properties.type === 'mosaic') {
+    saveableView = {
+      ...saveableView,
+      properties: {
+        ...saveableView.properties,
+        xColumn: getXColumnSelection(state),
+        ySeriesColumns: [getYColumnSelection(state)],
+        fillColumns: getFillColumnsSelection(state),
+      },
+    }
+  }
 
   if (saveableView.properties.type === 'histogram') {
     saveableView = {